# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phober - Device Service",
        "contact": {
            "name": "Hikmat",
            "url": "https://www.linkedin.com/in/abdukhaligov/",
            "email": "hikmat.pou@gmail.com"
        },
        "version": "V 1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8300",
            "description": "localhost"
        }
    ],
    "paths": {
        "/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get Games",
                "operationId": "games",
                "parameters": [
                    {
                        "name": "size",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get Game by ID",
                "operationId": "gameById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/games/search/by-device/{device}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get Games by Device slug",
                "operationId": "gamesByDevice",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "size",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/games/search/by-genre/{genre}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get Games by Genre slug",
                "operationId": "gamesByGenre",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "size",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/games/search/by-rating/{rating}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get Games by Rating",
                "operationId": "gamesByRating",
                "parameters": [
                    {
                        "name": "rating",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/games/search/multiplayer": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get Multiplayer Games",
                "operationId": "gamesMultiplayer",
                "parameters": [
                    {
                        "name": "size",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get Devices",
                "operationId": "devices",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        },
        "/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get Genres",
                "operationId": "genres",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": []
            }
        }
    }
}